// src/typings/models.ts
export enum AssetHistoryIntervals {
  M1 = "m1", // 1 minute
  M5 = "m5", // 5 minutes
  M15 = "m15", // 15 minutes
  M30 = "m30", // 30 minutes
  H1 = "h1", // 1 hour
  H2 = "h2", // 2 hours
  H6 = "h6", // 6 hours
  H12 = "h12", // 12 hours
  D1 = "d1", // 1 day
}

export enum AssetHistoryDuration {
  day = "1D",
  week = "1W",
  month = "1M",
  threeMonths = "3M",
  sixMonths = "6M",
  year = "1Y",
  all = "ALL",
}